Makes the name of the consequence operator a parameter
It is now a parameter of the Process_calculus module.

This allows for potentially multiple embeddings co-existing at the
same time, as long as they have a different consequence operator.

We need to make the names of the process definitions used in Cll_terms
also variable on the process calculus used.