Makes process definitions dependent on a process calculus prefix
This is the last hurdle to allow multiple simultaneous
correspondences.

This still does not allow multiple copies of the same
correspondence (though it is not clear why one would ever want this).

When using inference rules, process names (such as "TimesProc") must
be constant. Otherwise they get refreshed with the latest counter like
every other variable (e.g. "TimesProc8").

I feel there is a better way to do this than adding a prefix for each
correspondence, but this will require some rethinking of the embedding
and quite a bit of refactoring.