Stores provenance info in the process structure
Changes compose1, compose, and verify commands to not require an actionstate any
more.

Compose1 uses the action result as a label. This is rarely unique, but
we assume binary compositions take place in an independent context and
that a compose command will be issued at the end of the proof.

The compose command uses the final composition label in its state.

Closes #7